Move Price inner class from Callability to Bond.
