+++ ql/experimental/convertiblebonds/convertiblebond.hpp
@@ -41,7 +41,7 @@ namespace QuantLib {
     //! %callability leaving to the holder the possibility to convert
     class SoftCallability : public Callability {
       public:
-        SoftCallability(const Price& price,
+        SoftCallability(const Bond::Price& price,
                         const Date& date,
                         Real trigger)
         : Callability(price, Callability::Call, date), trigger_(trigger) {}
